Simplify implementation and remove obsolete code
