+++ servers/nextjs/app/(presentation-generator)/presentation/components/Chat.tsx
@@ -20,7 +20,6 @@ import React, {
   DragEvent,
   FormEvent,
   KeyboardEvent,
-  ReactNode,
   useCallback,
   useEffect,
   useRef,
@@ -32,23 +31,17 @@ import { useDispatch, useSelector } from "react-redux";
 import MarkdownRenderer from "@/components/MarkDownRender";
 import { ImagesApi } from "../../services/api/images";
 import { PresentationGenerationApi } from "../../services/api/presentation-generation";
-import { PresentationChatApi } from "../../services/api/chat";
 import {
   PRESENTON_BLANK_SLIDE_PROMPT_EVENT,
   type BlankSlidePromptEventDetail,
 } from "../../_shared/blank-slide-prompt-event";
 import type {
-  ChatAttachment,
   ChatConversationSummary,
   ChatHistoryMessage,
-  ChatMessageResponse,
-  ChatStreamHandlers,
   ChatStreamTrace,
 } from "../../services/api/chat";
 import ToolTip from "@/components/ToolTip";
 import { cn } from "@/lib/utils";
-import type { TemplateV2SurfaceSelectedDetail } from "@/components/slide-editor/events/events";
-import type { TemplateV2Layout } from "@/components/slide-editor/importing/template-v2-import";
 import {
   MAX_NUMBER_OF_SLIDES,
   MAX_OUTLINE_CONTENT_WORDS,
@@ -62,1101 +55,74 @@ import {
   PopoverTrigger,
 } from "@/components/ui/popover";
 import { LLM_PROVIDERS } from "@/utils/providerConstants";
-import type { LLMConfig } from "@/types/llm_config";
 import type { AppDispatch, RootState } from "@/store/store";
 import {
   setPresentationData,
   type PresentationData,
 } from "@/store/slices/presentationGeneration";
 
-const suggestions: { id: string; icon: ReactNode; suggestion: string }[] = [
-  {
-    id: "generate",
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="12"
-        height="12"
-        viewBox="0 0 12 12"
-        fill="none"
-        aria-hidden="true"
-      >
-        <g clipPath="url(#chat-suggestion-generate)">
-          <path
-            d="M10.82 1.82039L10.18 1.18039C10.1238 1.12355 10.0568 1.07842 9.98299 1.04763C9.90918 1.01683 9.83 1.00098 9.75002 1.00098C9.67005 1.00098 9.59087 1.01683 9.51706 1.04763C9.44325 1.07842 9.37628 1.12355 9.32002 1.18039L1.18002 9.32039C1.12318 9.37665 1.07806 9.44362 1.04726 9.51743C1.01647 9.59123 1.00061 9.67041 1.00061 9.75039C1.00061 9.83036 1.01647 9.90954 1.04726 9.98335C1.07806 10.0572 1.12318 10.1241 1.18002 10.1804L1.82002 10.8204C1.87593 10.8778 1.94279 10.9235 2.01664 10.9547C2.0905 10.9859 2.16985 11.0019 2.25002 11.0019C2.33019 11.0019 2.40955 10.9859 2.4834 10.9547C2.55726 10.9235 2.62411 10.8778 2.68002 10.8204L10.82 2.68039C10.8775 2.62448 10.9231 2.55762 10.9543 2.48377C10.9855 2.40991 11.0016 2.33056 11.0016 2.25039C11.0016 2.17022 10.9855 2.09087 10.9543 2.01701C10.9231 1.94316 10.8775 1.8763 10.82 1.82039Z"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M7 3.5L8.5 5"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M2.5 3V5"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M9.5 7V9"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M5 1V2"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M3.5 4H1.5"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M10.5 8H8.5"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M5.5 1.5H4.5"
-            stroke="#7F22FE"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </g>
-        <defs>
-          <clipPath id="chat-suggestion-generate">
-            <rect width="12" height="12" fill="white" />
-          </clipPath>
-        </defs>
-      </svg>
-    ),
-    suggestion: "Generate a full presentation from my topic",
-  },
-  {
-    id: "improve",
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="12"
-        height="12"
-        viewBox="0 0 12 12"
-        fill="none"
-        aria-hidden="true"
-      >
-        <g clipPath="url(#chat-suggestion-improve)">
-          <path
-            d="M4.96847 7.75012C4.92383 7.57709 4.83364 7.41918 4.70728 7.29282C4.58092 7.16646 4.42301 7.07626 4.24997 7.03162L1.18247 6.24062C1.13014 6.22577 1.08407 6.19425 1.05128 6.15085C1.01848 6.10744 1.00073 6.05453 1.00073 6.00012C1.00073 5.94572 1.01848 5.89281 1.05128 5.8494C1.08407 5.806 1.13014 5.77448 1.18247 5.75962L4.24997 4.96812C4.42294 4.92353 4.58082 4.83341 4.70717 4.70714C4.83353 4.58088 4.92375 4.42307 4.96847 4.25012L5.75947 1.18262C5.77417 1.13008 5.80566 1.0838 5.84913 1.05082C5.8926 1.01785 5.94566 1 6.00022 1C6.05478 1 6.10784 1.01785 6.15131 1.05082C6.19478 1.0838 6.22627 1.13008 6.24097 1.18262L7.03147 4.25012C7.07611 4.42316 7.1663 4.58107 7.29266 4.70743C7.41902 4.83379 7.57693 4.92399 7.74997 4.96862L10.8175 5.75912C10.8702 5.77367 10.9167 5.80513 10.9499 5.84866C10.983 5.8922 11.001 5.94541 11.001 6.00012C11.001 6.05484 10.983 6.10805 10.9499 6.15159C10.9167 6.19512 10.8702 6.22657 10.8175 6.24112L7.74997 7.03162C7.57693 7.07626 7.41902 7.16646 7.29266 7.29282C7.1663 7.41918 7.07611 7.57709 7.03147 7.75012L6.24047 10.8176C6.22577 10.8702 6.19428 10.9165 6.15081 10.9494C6.10734 10.9824 6.05428 11.0002 5.99972 11.0002C5.94516 11.0002 5.8921 10.9824 5.84863 10.9494C5.80516 10.9165 5.77367 10.8702 5.75897 10.8176L4.96847 7.75012Z"
-            stroke="#155DFC"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M10 1.5V3.5"
-            stroke="#155DFC"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M11 2.5H9"
-            stroke="#155DFC"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M2 8.5V9.5"
-            stroke="#155DFC"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M2.5 9H1.5"
-            stroke="#155DFC"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </g>
-        <defs>
-          <clipPath id="chat-suggestion-improve">
-            <rect width="12" height="12" fill="white" />
-          </clipPath>
-        </defs>
-      </svg>
-    ),
-    suggestion: "Improve this slide content",
-  },
-  {
-    id: "rewrite",
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="12"
-        height="12"
-        viewBox="0 0 12 12"
-        fill="none"
-        aria-hidden="true"
-      >
-        <path
-          d="M6 10H10.5"
-          stroke="#009966"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M8.18799 1.81087C8.38703 1.61182 8.657 1.5 8.93849 1.5C9.21998 1.5 9.48994 1.61182 9.68899 1.81087C9.88803 2.00991 9.99986 2.27988 9.99986 2.56137C9.99986 2.84286 9.88803 3.11282 9.68899 3.31187L3.68399 9.31737C3.56504 9.43632 3.418 9.52333 3.25649 9.57037L1.82049 9.98937C1.77746 10.0019 1.73186 10.0027 1.68844 9.99155C1.64503 9.98042 1.6054 9.95783 1.57371 9.92614C1.54202 9.89445 1.51943 9.85483 1.50831 9.81141C1.49719 9.768 1.49794 9.72239 1.51049 9.67937L1.92949 8.24337C1.9766 8.08203 2.06361 7.93518 2.18249 7.81637L8.18799 1.81087Z"
-          stroke="#009966"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    ),
-    suggestion: "Rewrite this content professionally",
-  },
-  {
-    id: "notes",
-    icon: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="12"
-        height="12"
-        viewBox="0 0 12 12"
-        fill="none"
-        aria-hidden="true"
-      >
-        <path
-          d="M1.5 1.5V9.5C1.5 9.76522 1.60536 10.0196 1.79289 10.2071C1.98043 10.3946 2.23478 10.5 2.5 10.5H10.5"
-          stroke="#E17100"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M9 8.5V4.5"
-          stroke="#E17100"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M6.5 8.5V2.5"
-          stroke="#E17100"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M4 8.5V7"
-          stroke="#E17100"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    ),
-    suggestion: "Add speaker notes to this slide",
-  },
-];
-
-const outlineQuickPrompts = [
-  "Expand outline",
-  "Shorten outline",
-  "Reorder sections",
-  "Merge similar slides",
-  "Split large sections",
-  "Improve conclusion",
-  "Improve introduction",
-];
-
-const presentationQuickPrompts = [
-  "Create an executive summary",
-  "Strengthen the story flow",
-  "Add data and citations",
-  "Create speaker notes",
-];
-
-const templateV2QuickPrompts = [
-  "Improve this slide's layout",
-  "Rewrite this slide for executives",
-  "Add a supporting visual",
-  "Make the deck visually consistent",
-  "Add data and source citations",
-  "Create speaker notes for this slide",
-];
-
-const editorQuickPrompts = [
-  "Rewrite for executives",
-  "Improve slide layout",
-  "Add data & citations",
-  "Create speaker notes",
-  "Make the deck consistent",
-];
-
-const outlineEditorQuickPrompts = [
-  "Strengthen the story flow",
-  "Make the outline concise",
-  "Reorder the sections",
-  "Merge similar slides",
-  "Improve the conclusion",
-];
-
-const quickPromptGroups = [
-  {
-    label: "Popular",
-    prompts: ["Make it shorter", "Make this smaller", "Generate a new image and replace this one"],
-  },
-  {
-    label: "Add Data",
-    prompts: ["Add data and citations", "Add a chart", "Add a table"],
-  },
-  {
-    label: "Add Visuals",
-    prompts: ["Generate a new image and replace this one", "Add a chart", "Add a table"],
-  },
-];
-
-const outlineQuickPromptGroups = [
-  {
-    label: "Popular",
-    prompts: [
-      "Make the outline shorter",
-      "Expand the outline",
-      "Strengthen the story flow",
-    ],
-  },
-  {
-    label: "Structure",
-    prompts: [
-      "Reorder the sections",
-      "Merge similar slides",
-      "Split large sections",
-    ],
-  },
-  {
-    label: "Refine Content",
-    prompts: [
-      "Rewrite for executives",
-      "Improve the introduction",
-      "Improve the conclusion",
-    ],
-  },
-];
-
-const getSelectedTextModel = (config: LLMConfig) => {
-  switch (config.LLM) {
-    case "openai":
-      return config.OPENAI_MODEL;
-    case "deepseek":
-      return config.DEEPSEEK_MODEL;
-    case "google":
-      return config.GOOGLE_MODEL;
-    case "vertex":
-      return config.VERTEX_MODEL;
-    case "azure":
-      return config.AZURE_OPENAI_MODEL;
-    case "bedrock":
-      return config.BEDROCK_MODEL;
-    case "openrouter":
-      return config.OPENROUTER_MODEL;
-    case "fireworks":
-      return config.FIREWORKS_MODEL;
-    case "together":
-      return config.TOGETHER_MODEL;
-    case "cerebras":
-      return config.CEREBRAS_MODEL;
-    case "litellm":
-      return config.LITELLM_MODEL;
-    case "lmstudio":
-      return config.LMSTUDIO_MODEL;
-    case "anthropic":
-      return config.ANTHROPIC_MODEL;
-    case "ollama":
-      return config.OLLAMA_MODEL;
-    case "custom":
-      return config.CUSTOM_MODEL;
-    case "codex":
-      return config.CODEX_MODEL;
-    default:
-      return undefined;
-  }
-};
-
-type ChatMessage = {
-  id: string;
-  role: "user" | "assistant" | "error";
-  content: string;
-  toolCalls?: string[];
-  activity?: AssistantActivity[];
-  layoutPreview?: ChatLayoutPreview;
-  editPreview?: ChatEditPreview;
-};
-
-type ChatEditPreview = {
-  originalSlides: unknown[];
-  modifiedSlides?: unknown[];
-  slideIndices: number[];
-  changeCount: number;
-};
-
-type ChatLayoutPreview = {
-  layout: TemplateV2Layout;
-  layoutId?: string | null;
-  slideIndex?: number | null;
-};
-
-type SubmitMessageOptions = {
-  backendContext?: string;
-  layoutPreview?: ChatLayoutPreview;
-};
-
-type PastedChatImage = {
-  id: string;
-  name: string;
-  url: string;
-  file?: File;
-  extractedText?: string;
-};
-
-type ChatLink = {
-  id: string;
-  url: string;
-};
-
-type ChatDocumentAttachment = {
-  id: string;
-  name: string;
-  filePath: string;
-  mimeType?: string;
-};
-
-type ChatProps = {
-  presentationId: string;
-  presentationData?: unknown;
-  resourceId?: string;
-  chatAdapter?: ChatApiAdapter;
-  conversationStorageScope?: string;
-  resourceLabel?: string;
-  variant?: "presentation" | "outline" | "template-v2";
-  useEditorLayout?: boolean;
-  inputDisabled?: boolean;
-  currentSlide?: number;
-  selectedTemplateV2Target?: TemplateV2SurfaceSelectedDetail["selection"];
-  onClearChatSlideReference?: () => void;
-  onClearChatTargetReference?: () => void;
-  onBeforeSend?: () => Promise<void> | void;
-  onPresentationChanged?: () => Promise<void> | void;
-  onChatMutationStateChange?: (isMutating: boolean) => void;
-  onAgentSlideFocus?: (focus: {
-    slideIndex: number;
-    eventId: string;
-    tool?: string;
-    status?: string;
-    isMutatingTool: boolean;
-  }) => void;
-  onChatSendingStateChange?: (isSending: boolean) => void;
-  onFollowModeChange?: (isEnabled: boolean) => void;
-};
-
-export type ChatApiAdapter = {
-  listConversations: (resourceId: string) => Promise<ChatConversationSummary[]>;
-  getHistory: (
-    resourceId: string,
-    conversationId: string
-  ) => Promise<{ messages: ChatHistoryMessage[] }>;
-  deleteConversation: (
-    resourceId: string,
-    conversationId: string
-  ) => Promise<void>;
-  streamMessage: (
-    payload: {
-      resourceId: string;
-      message: string;
-      conversation_id?: string;
-      attachments?: ChatAttachment[];
-    },
-    handlers?: ChatStreamHandlers,
-    options?: { signal?: AbortSignal }
-  ) => Promise<ChatMessageResponse>;
-};
-
-const presentationChatAdapter: ChatApiAdapter = {
-  listConversations: (resourceId) =>
-    PresentationChatApi.listConversations(resourceId),
-  getHistory: (resourceId, conversationId) =>
-    PresentationChatApi.getHistory(resourceId, conversationId),
-  deleteConversation: (resourceId, conversationId) =>
-    PresentationChatApi.deleteConversation(resourceId, conversationId),
-  streamMessage: (payload, handlers, options) =>
-    PresentationChatApi.streamMessage(
-      {
-        presentation_id: payload.resourceId,
-        message: payload.message,
-        conversation_id: payload.conversation_id,
-        attachments: payload.attachments,
-      },
-      handlers,
-      options
-    ),
-};
-
-type AssistantActivity = {
-  id: string;
-  label: string;
-  kind?: string;
-  round?: number;
-  tool?: string;
-  state: "running" | "success" | "error" | "info";
-};
-
-type AssistantPromptMetrics = {
-  startedAt: number;
-  attachmentImageCount: number;
-  attachmentDocumentCount: number;
-  linkCount: number;
-  mutatingToolCount: number;
-  readToolCount: number;
-  uniqueTools: Set<string>;
-  mutatedSlides: Set<number>;
-};
-
-const createMessageId = () => {
-  if (typeof crypto !== "undefined" && "randomUUID" in crypto) {
-    return crypto.randomUUID();
-  }
-
-  return `${Date.now()}-${Math.random().toString(16).slice(2)}`;
-};
-
-const createChatLayoutPreviewSlide = (preview: ChatLayoutPreview) => ({
-  id: `chat-layout-preview-${preview.slideIndex ?? "slide"}`,
-  content: {},
-  ui: preview.layout,
-  layout: preview.layoutId || "chat-layout-preview",
-  layout_group: "template-v2",
-});
-
-const getPresentationSlide = (presentationData: unknown, slideIndex: number) => {
-  if (!presentationData || typeof presentationData !== "object") return null;
-  const slides = (presentationData as Record<string, unknown>).slides;
-  return Array.isArray(slides) ? slides[slideIndex] ?? null : null;
-};
-
-const getPresentationFonts = (presentationData: unknown) => {
-  if (!presentationData || typeof presentationData !== "object") return undefined;
-  return (presentationData as Record<string, unknown>).fonts;
-};
-
-const clonePreviewSlide = (slide: unknown) => {
-  if (!slide) return null;
-  try {
-    return structuredClone(slide);
-  } catch {
-    try {
-      return JSON.parse(JSON.stringify(slide)) as unknown;
-    } catch {
-      return null;
-    }
-  }
-};
-
-const URL_PATTERN =
-  /(https?:\/\/[^\s<>"']+\.[^\s<>"']+|www\.[^\s<>"']+\.[^\s<>"']+)/gi;
-const IMAGE_READ_INTENT_PATTERN =
-  /\b(read|extract|parse|analy[sz]e|summari[sz]e|ocr|text|table|chart|data|numbers?|metrics?)\b/i;
-const IMAGE_EXTENSION_PATTERN = /\.(jpe?g|png|gif|bmp|tiff?|webp)$/i;
-const ATTACHMENT_CONTENT_LIMIT = 2000;
-
-function pullLinksFromText(text: string) {
-  const links: ChatLink[] = [];
-  const cleanText = text
-    .replace(URL_PATTERN, (match) => {
-      const url = match.replace(/[.,;:!?)}\]]+$/g, "");
-      links.push({
-        id: createMessageId(),
-        url: url.startsWith("www.") ? `https://${url}` : url,
-      });
-      return match.slice(url.length);
-    });
-  return { cleanText, links };
-}
-
-function appendInputText(previous: string, next: string) {
-  if (!next) return previous;
-  if (!previous) return next.trimStart();
-  if (/\s$/.test(previous) || /^\s/.test(next)) return `${previous}${next}`;
-  return `${previous} ${next}`;
-}
-
-function isImageFile(file: File) {
-  return (
-    file.type.startsWith("image/") || IMAGE_EXTENSION_PATTERN.test(file.name)
-  );
-}
-
-function shouldReadAttachedImages(message: string) {
-  return IMAGE_READ_INTENT_PATTERN.test(message);
-}
-
-function trimAttachmentContent(content: string) {
-  if (content.length <= ATTACHMENT_CONTENT_LIMIT) return content;
-  return `${content.slice(0, ATTACHMENT_CONTENT_LIMIT)}\n[Attachment truncated]`;
-}
-
-function buildChatDocumentAttachments(
-  documents: ChatDocumentAttachment[]
-): ChatAttachment[] {
-  return documents.map((document) => ({
-    type: "document",
-    name: document.name,
-    file_path: document.filePath,
-    mime_type: document.mimeType || null,
-  }));
-}
-
-function hasDraggedFiles(event: DragEvent<HTMLElement>) {
-  return (
-    Array.from(event.dataTransfer.types ?? []).includes("Files") ||
-    event.dataTransfer.files.length > 0 ||
-    Array.from(event.dataTransfer.items ?? []).some(
-      (item) => item.kind === "file"
-    )
-  );
-}
-
-function getDroppedFileUri(event: DragEvent<HTMLElement>) {
-  if (!Array.from(event.dataTransfer.types ?? []).includes("text/uri-list")) {
-    return "";
-  }
-  return event.dataTransfer.getData("text/uri-list");
-}
-
-async function readDecomposedFile(filePath: string) {
-  if (typeof window !== "undefined" && window.electron?.readFile) {
-    const result = await window.electron.readFile(filePath);
-    return typeof result === "string" ? result : result?.content || "";
-  }
-
-  const response = await fetch("/api/read-file", {
-    method: "POST",
-    body: JSON.stringify({ filePath }),
-  });
-  const result = await response.json();
-  if (!response.ok) {
-    throw new Error(result?.error || "Failed to read document.");
-  }
-  return result?.content || "";
-}
-
-const conversationStorageKey = (scope: string, resourceId: string) =>
-  `presenton:chat:${scope}:conversationId:${resourceId}`;
-
-const readStoredConversationId = (key: string) => {
-  if (typeof window === "undefined") return null;
-  try {
-    return window.sessionStorage.getItem(key);
-  } catch {
-    return null;
-  }
-};
-
-const storeConversationId = (key: string, conversationId: string) => {
-  if (typeof window === "undefined") return;
-  try {
-    window.sessionStorage.setItem(key, conversationId);
-  } catch {
-    // Chat history still works from the server when browser storage is blocked.
-  }
-};
-
-const removeStoredConversationId = (key: string) => {
-  if (typeof window === "undefined") return;
-  try {
-    window.sessionStorage.removeItem(key);
-  } catch {
-    // Nothing else needs to be cleared when browser storage is unavailable.
-  }
-};
-
-const AssistantSparkleIcon = ({ size = 14 }: { size?: number }) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 14 14"
-    fill="none"
-    aria-hidden="true"
-  >
-    <path
-      d="M6.9 1.1c.22 3.03 1.37 4.17 4.4 4.4-3.03.22-4.18 1.37-4.4 4.4-.22-3.03-1.37-4.18-4.4-4.4 3.03-.23 4.18-1.37 4.4-4.4Z"
-      fill="#7A5AF8"
-    />
-    <path
-      d="M11.2 9.4c.1 1.42.64 1.96 2.06 2.06-1.42.1-1.96.64-2.06 2.06-.1-1.42-.64-1.96-2.06-2.06 1.42-.1 1.96-.64 2.06-2.06Z"
-      fill="#6172F3"
-    />
-  </svg>
-);
-
-const AssistantMarker = () => (
-  <div className="mb-2 flex items-center gap-1.5 text-[#8A8F98]">
-    <AssistantSparkleIcon size={14} />
-    <span className="text-[11px] font-medium leading-4">Assistant</span>
-  </div>
-);
-
-const ActivityStatusIcon = ({ activity }: { activity: AssistantActivity }) => {
-  if (activity.state === "running") {
-    return (
-      <span
-        className="activity-flow-dots mt-1 relative h-[9px] w-[22px] shrink-0"
-        aria-label="Working"
-      >
-        <span className="absolute left-0 top-[1.5px] h-[6px] w-[6px] rounded-full bg-[#C3C3CB]" />
-        <span className="absolute left-[8px] top-[1.5px] h-[6px] w-[6px] rounded-full bg-[#C3C3CB]" />
-        <span className="absolute left-[16px] top-[1.5px] h-[6px] w-[6px] rounded-full bg-[#C3C3CB]" />
-      </span>
-    );
-  }
-
-  if (activity.state === "error") {
-    return <span className="h-2 w-2 shrink-0 rounded-full bg-[#F04438]" />;
-  }
-
-  return <span className="h-2 w-2 shrink-0 rounded-full bg-[#E6E6E6]" />;
-};
-
-const QuickPromptsPanel = ({
-  onPromptSelect,
-  groups = quickPromptGroups,
-}: {
-  onPromptSelect: (prompt: string) => void;
-  groups?: typeof quickPromptGroups;
-}) => (
-  <div className="flex flex-col gap-6 font-syne">
-    <AssistantSparkleIcon size={24} />
-    <div className="flex flex-col gap-5">
-      {groups.map((group) => (
-        <div key={group.label} className="flex flex-col gap-2">
-          <p className="text-sm font-normal tracking-[0.28px] text-[#333333]">
-            {group.label}
-          </p>
-          <div className="flex flex-wrap gap-1.5">
-            {group.prompts.map((prompt) => (
-              <button
-                key={prompt}
-                type="button"
-                onClick={() => onPromptSelect(prompt)}
-                className="rounded-full border border-black/[0.04] bg-black/[0.02] px-4 py-1.5 font-manrope text-xs font-normal tracking-[0.24px] text-[#333333] outline-none transition-colors hover:border-[#D9D6FE] hover:bg-[#FAFAFF] hover:text-[#7A5AF8] focus:border-[#7A5AF8] focus:outline-none focus:ring-0 focus-visible:border-[#7A5AF8] focus-visible:text-[#7A5AF8] focus-visible:outline-none focus-visible:ring-0 active:border-[#7A5AF8] active:text-[#7A5AF8]"
-              >
-                {prompt}
-              </button>
-            ))}
-          </div>
-        </div>
-      ))}
-    </div>
-  </div>
-);
-
-const EditComparisonPreview = ({
-  preview,
-  fonts,
-  selectedVersion,
-  isApplying,
-  onSelectVersion,
-}: {
-  preview: ChatEditPreview;
-  fonts?: unknown;
-  selectedVersion: "original" | "modified";
-  isApplying: boolean;
-  onSelectVersion: (version: "original" | "modified") => void;
-}) => {
-  if (!preview.modifiedSlides?.length) return null;
-
-  const cards = [
-    { label: "Original", slides: preview.originalSlides, version: "original" as const },
-    { label: "Modified", slides: preview.modifiedSlides, version: "modified" as const },
-  ];
-
-  return (
-    <div className="flex w-full flex-col gap-2 font-manrope">
-      <div className="flex items-center gap-1 text-[13px] leading-[18px]">
-        <Image
-          src="/frame.svg"
-          alt=""
-          width={14}
-          height={14}
-          className="h-[14px] w-[14px] shrink-0"
-        />
-        <span className="font-semibold text-[#191919]">Select edits</span>
-        <span className="ml-auto text-[11px] font-medium leading-[normal] text-[#7A5AF8]">
-          {preview.changeCount} {preview.changeCount === 1 ? "Change" : "Changes"}
-        </span>
-      </div>
-      <div className="grid grid-cols-2 gap-[5px]">
-        {cards.map((card) => (
-          <button
-            key={card.label}
-            type="button"
-            onClick={() => onSelectVersion(card.version)}
-            disabled={isApplying}
-            className={cn(
-              "min-w-0 overflow-hidden rounded-[6px] border bg-[#F9FAFB] px-[6px] py-[10px] text-left transition-[border-color,background-color,box-shadow,opacity] hover:border-[#B7ACFC] disabled:cursor-wait disabled:opacity-70",
-              selectedVersion === card.version
-                ? "border-[#7A5AF8] bg-[#FAFAFF]"
-                : "border-[#EDEEEF]",
-            )}
-            aria-pressed={selectedVersion === card.version}
-            aria-label={`Restore ${card.label.toLowerCase()} slide state`}
-          >
-            <span className="mb-[7px] flex items-center justify-center gap-1 truncate text-center text-[13px] font-medium leading-[normal] text-[#191919]">
-              {isApplying && selectedVersion === card.version && (
-                <Loader2 className="h-3 w-3 animate-spin text-[#7A5AF8]" />
-              )}
-              <span>{card.label}</span>
-            </span>
-            <span className="flex flex-col gap-[3px]">
-              {card.slides.slice(0, 2).map((slide, index) => (
-                <TemplateV2HtmlSlidePreview
-                  key={`${card.label}-${index}`}
-                  slide={slide}
-                  fonts={fonts}
-                  className="overflow-hidden rounded-[2px] border border-[#EDEEEF] bg-white"
-                />
-              ))}
-            </span>
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const TOOL_LABELS: Record<string, string> = {
-  addOutline: "Outline adder",
-  updateOutline: "Outline editor",
-  deleteOutline: "Outline remover",
-  addNewSlide: "Blank slide adder",
-  addNewSlideLayout: "Layout slide adder",
-  getAvailableLayouts: "Layout finder",
-  getTemplateSummary: "Template reader",
-  readSourceDocuments: "Source document reader",
-  searchSlide: "Slide search",
-  getSlideAtIndex: "Slide reader",
-  saveSlide: "Slide saver",
-  updateSlide: "Slide updater",
-  deleteSlide: "Slide remover",
-  addElement: "Element adder",
-  updateElement: "Element updater",
-  deleteElement: "Element remover",
-  addComponent: "Component adder",
-  createComponent: "Component creator",
-  updateComponent: "Component updater",
-  deleteComponent: "Component remover",
-  getPresentationTheme: "Theme reader",
-  setPresentationTheme: "Theme applier",
-  generateAssets: "Asset generator",
-};
-
-const MUTATING_TOOLS = new Set([
-  "addOutline",
-  "updateOutline",
-  "deleteOutline",
-  "addNewSlide",
-  "addNewSlideLayout",
-  "saveSlide",
-  "updateSlide",
-  "deleteSlide",
-  "addElement",
-  "updateElement",
-  "deleteElement",
-  "addComponent",
-  "createComponent",
-  "updateComponent",
-  "deleteComponent",
-  "setPresentationTheme",
-]);
-// Only focus slides when the agent is actively mutating them.
-// Read/open traces (e.g. getSlideAtIndex) can happen ahead of edits and feel jumpy.
-const SLIDE_FOCUS_TOOLS = new Set([
-  "addNewSlide",
-  "addNewSlideLayout",
-  "saveSlide",
-  "updateSlide",
-  "deleteSlide",
-  "addElement",
-  "updateElement",
-  "deleteElement",
-  "addComponent",
-  "createComponent",
-  "updateComponent",
-  "deleteComponent",
-]);
-const SLIDE_FOCUS_STATUSES = new Set(["start"]);
-const MIN_SLIDE_FOCUS_DWELL_MS = 700;
-
-const getToolLabel = (tool?: string) => {
-  if (!tool) {
-    return "";
-  }
-  return TOOL_LABELS[tool] ?? tool;
-};
-
-const humanizeTraceMessage = (message: string, tool?: string) => {
-  const trimmed = message.trim();
-  if (!trimmed) {
-    return "";
-  }
-
-  const lower = trimmed.toLowerCase();
-  if (lower === "reading deck context") {
-    return "Reviewing your presentation context.";
-  }
-  if (lower === "reading the presentation outline") {
-    return "Reading the presentation outline.";
-  }
-  if (lower === "reading the outline draft") {
-    return "Reading the outline draft.";
-  }
-  if (lower === "adding an outline slide") {
-    return "Adding an outline slide.";
-  }
-  if (lower === "updating the outline slide") {
-    return "Updating the outline slide.";
-  }
-  if (lower === "deleting the outline slide") {
-    return "Deleting the outline slide.";
-  }
-  if (lower === "reordering outline slides") {
-    return "Reordering outline slides.";
-  }
-  if (lower === "searching relevant slides") {
-    return "Searching slides for relevant content.";
-  }
-  if (lower === "opening the requested slide") {
-    return "Opening the selected slide.";
-  }
-  if (lower === "checking available themes") {
-    return "Checking available color themes.";
-  }
-  if (lower === "checking available layouts") {
-    return "Checking available layouts.";
-  }
-  if (lower === "checking the layout schema") {
-    return "Validating the slide schema.";
-  }
-  if (lower === "generating slide assets") {
-    return "Generating images and icons.";
-  }
-  if (lower === "saving the slide") {
-    return "Saving slide updates.";
-  }
-  if (lower === "deleting the slide") {
-    return "Deleting the slide.";
-  }
-  if (lower === "applying presentation theme") {
-    return "Applying the selected theme.";
-  }
-  if (lower === "reading template structure") {
-    return "Reading the template structure.";
-  }
-  if (lower === "reading source documents") {
-    return "Reading the source documents.";
-  }
-  if (lower === "opening the requested template slide") {
-    return "Opening the selected template slide.";
-  }
-  if (lower === "searching template content") {
-    return "Searching template content.";
-  }
-  if (lower === "finding editable elements") {
-    return "Finding editable elements.";
-  }
-  if (lower === "updating template content") {
-    return "Updating template content.";
-  }
-  if (lower === "deleting the template component") {
-    return "Deleting the selected component.";
-  }
-  if (lower === "swapping component variant") {
-    return "Swapping the component variant.";
-  }
-  if (lower.startsWith("using tools:")) {
-    const toolNames = trimmed
-      .slice("using tools:".length)
-      .split(",")
-      .map((entry) => entry.trim())
-      .filter(Boolean)
-      .map((entry) => getToolLabel(entry));
-    return toolNames.length === 0
-      ? "Planning the next step."
-      : "Choosing the best way to help.";
-  }
-  if (lower.includes("found requested data")) {
-    if (tool === "getSlideAtIndex") {
-      return "Found the requested slide details.";
-    }
-    return "Found the requested information.";
-  }
-  if (lower.endsWith("completed.")) {
-    return trimmed;
-  }
-  if (lower.includes("failed")) {
-    return trimmed;
-  }
-  return trimmed;
-};
-
-const inferStatusState = (status: string): AssistantActivity["state"] => {
-  const normalized = status.trim().toLowerCase();
-  if (
-    normalized.includes("preparing") ||
-    normalized.includes("thinking") ||
-    normalized.includes("reading") ||
-    normalized.includes("searching") ||
-    normalized.includes("opening") ||
-    normalized.includes("generating") ||
-    normalized.includes("processing") ||
-    normalized.includes("finalizing") ||
-    normalized.includes("saving")
-  ) {
-    return "running";
-  }
-
-  return "info";
-};
-
-const isAbortError = (error: unknown) =>
-  (error instanceof DOMException && error.name === "AbortError") ||
-  (error instanceof Error &&
-    error.message.toLowerCase().includes("aborted") &&
-    error.message.toLowerCase().includes("request"));
-
-const stripBackendContextFromUserMessage = (rawMessage: string) => {
-  const message = rawMessage ?? "";
-  if (!message.startsWith("UI context:")) {
-    return message;
-  }
-
-  const marker = "\nUser message:";
-  const markerIndex = message.indexOf(marker);
-  if (markerIndex === -1) {
-    return message;
-  }
-
-  return message.slice(markerIndex + marker.length).trimStart();
-};
-
-const formatTraceActivity = (
-  trace: ChatStreamTrace
-): Omit<AssistantActivity, "id"> | null => {
-  if (typeof trace.message === "string" && trace.message.trim().length > 0) {
-    return {
-      label: humanizeTraceMessage(trace.message, trace.tool),
-      kind: trace.kind,
-      round: trace.round,
-      tool: trace.tool,
-      state:
-        trace.status === "error"
-          ? "error"
-          : trace.status === "success"
-            ? "success"
-            : trace.status === "ready" || trace.status === "info"
-              ? "info"
-              : "running",
-    };
-  }
-
-  if (trace.tool && trace.status === "start") {
-    return {
-      label: humanActivityForTool(trace.tool, "start"),
-      kind: trace.kind,
-      round: trace.round,
-      tool: trace.tool,
-      state: "running",
-    };
-  }
-
-  if (trace.tool && trace.status === "success") {
-    return {
-      label: humanActivityForTool(trace.tool, "success"),
-      kind: trace.kind,
-      round: trace.round,
-      tool: trace.tool,
-      state: "success",
-    };
-  }
-
-  if (trace.tool && trace.status === "error") {
-    return {
-      label: "I could not finish that step.",
-      kind: trace.kind,
-      round: trace.round,
-      tool: trace.tool,
-      state: "error",
-    };
-  }
-
-  if (
-    trace.kind === "tool_plan" &&
-    Array.isArray(trace.tools) &&
-    trace.tools.length
-  ) {
-    return {
-      label: "Planning the next step.",
-      kind: trace.kind,
-      round: trace.round,
-      state: "info",
-    };
-  }
-
-  return null;
-};
-
-const humanActivityForTool = (
-  tool: string | undefined,
-  state: "start" | "success"
-) => {
-  const isDone = state === "success";
-  switch (tool) {
-    case "searchSlide":
-      return isDone ? "Found the relevant content." : "Looking through the content.";
-    case "getSlideAtIndex":
-      return isDone ? "Checked the slide." : "Checking the slide.";
-    case "addNewSlide":
-    case "addNewSlideLayout":
-    case "updateElement":
-    case "updateComponent":
-    case "addElement":
-    case "addComponent":
-    case "createComponent":
-    case "updateSlide":
-    case "saveSlide":
-      return isDone ? "Applied the change." : "Applying the change.";
-    case "deleteComponent":
-    case "deleteElement":
-    case "deleteSlide":
-      return isDone ? "Removed the selected item." : "Removing the selected item.";
-    case "generateAssets":
-      return isDone ? "Prepared the visual assets." : "Preparing visual assets.";
-    case "setPresentationTheme":
-      return isDone ? "Updated the theme." : "Updating the theme.";
-    default:
-      return isDone ? "Finished that step." : "Working on it.";
-  }
-};
-
-const readTraceSlideIndex = (trace: ChatStreamTrace) => {
-  if (typeof trace.slideIndex === "number" && trace.slideIndex >= 0) {
-    return trace.slideIndex;
-  }
-  if (typeof trace.slideNumber === "number" && trace.slideNumber > 0) {
-    return trace.slideNumber - 1;
-  }
-  return null;
-};
-
+import {
+  editorQuickPrompts,
+  getSelectedTextModel,
+  outlineEditorQuickPrompts,
+  outlineQuickPromptGroups,
+  outlineQuickPrompts,
+  presentationQuickPrompts,
+  quickPromptGroups,
+  suggestions,
+  templateV2QuickPrompts,
+} from "./chat/chat-prompts";
+import { presentationChatAdapter } from "./chat/chat-adapter";
+import type {
+  AssistantActivity,
+  AssistantPromptMetrics,
+  ChatDocumentAttachment,
+  ChatEditPreview,
+  ChatLink,
+  ChatMessage,
+  ChatProps,
+  PastedChatImage,
+  SubmitMessageOptions,
+} from "./chat/chat-types";
+import {
+  appendInputText,
+  buildChatDocumentAttachments,
+  clonePreviewSlide,
+  conversationStorageKey,
+  createChatLayoutPreviewSlide,
+  createMessageId,
+  getDroppedFileUri,
+  getPresentationFonts,
+  getPresentationSlide,
+  hasDraggedFiles,
+  isImageFile,
+  pullLinksFromText,
+  readDecomposedFile,
+  readStoredConversationId,
+  removeStoredConversationId,
+  shouldReadAttachedImages,
+  storeConversationId,
+  trimAttachmentContent,
+} from "./chat/chat-utils";
+import {
+  formatTraceActivity,
+  inferStatusState,
+  isAbortError,
+  MIN_SLIDE_FOCUS_DWELL_MS,
+  MUTATING_TOOLS,
+  readTraceSlideIndex,
+  SLIDE_FOCUS_STATUSES,
+  SLIDE_FOCUS_TOOLS,
+  stripBackendContextFromUserMessage,
+} from "./chat/chat-activity";
+import {
+  ActivityStatusIcon,
+  AssistantMarker,
+  EditComparisonPreview,
+  QuickPromptsPanel,
+} from "./chat/chat-widgets";
+
+export type { ChatApiAdapter } from "./chat/chat-types";
 const Chat = ({
   presentationId,
   presentationData,
